requirements.txt
Django==4.1.5
djangorestframework~=3.14.0
django-environ~=0.9.0
djangorestframework-simplejwt~=5.2.2
drf-yasg~=1.21.4
